chore: reorganizing the files on the frontend directory
