+++ frontend/src/app/createList/(components)/NewListForm/NewList.tsx
@@ -2,7 +2,7 @@
 
 import { useContextWrap } from '@/contextAPI/context'
 import checkAuth from '@/functions/checkAuthFunction'
-import useSubmitList from '@/functions/useSubmitList'
+import useSubmitList from '@/app/createList/(hooks)/useSubmitList'
 import newListSchema from '@/zodSchemas/newListSchema'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { SubmitHandler, useFieldArray, useForm } from 'react-hook-form'
