+++ frontend/src/app/dashboard/(hooks)/useDeleteWedding.ts
