+++ frontend/src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import LoggedHeader from '../../components/Headers/LoggedHeader'
-import WeddingsOwn from '../../components/weddingsDisplay/OwnWeddings'
-import WeddingsGuest from '../../components/weddingsDisplay/GuestWeddings'
+import WeddingsOwn from './(components)/weddingsDisplay/OwnWeddings'
+import WeddingsGuest from './(components)/weddingsDisplay/GuestWeddings'
 
 export default function Dashboard() {
   return (
