+++ frontend/src/app/giftList/[id]/(components)/conditionalRender/conditionalRender.tsx
@@ -1,9 +1,9 @@
 'use client'
-import GuestList from '@/components/giftsListDisplay/GuestList'
-import GuestRequest from '@/components/giftsListDisplay/GuestRequest'
-import OwnerList from '@/components/giftsListDisplay/OwnerList'
+import GuestList from '@/app/giftList/[id]/(components)/giftsListDisplay/GuestList'
+import GuestRequest from '@/app/giftList/[id]/(components)/giftsListDisplay/GuestRequest'
+import OwnerList from '@/app/giftList/[id]/(components)/giftsListDisplay/OwnerList'
 import { useContextWrap } from '@/contextAPI/context'
-import useGetData from '@/functions/getWeddingDataFunction'
+import useGetData from '@/app/giftList/[id]/(hooks)/useGetData'
 import { useEffect } from 'react'
 
 export default function ConditionalRenderingListPage() {
