+++ frontend/src/app/giftList/[id]/(hooks)/useDeleteGift.ts
