+++ frontend/src/app/giftList/[id]/(hooks)/useSubmitNewGifts.ts
