+++ frontend/src/app/giftList/[id]/page.tsx
@@ -1,6 +1,6 @@
 import LoggedHeader from '../../../components/Headers/LoggedHeader'
 
-import ConditionalRenderingListPage from './(components)/conditionalRender'
+import ConditionalRenderingListPage from './(components)/conditionalRender/conditionalRender'
 
 export default function giftsList() {
   return (
