+++ frontend/src/app/registerPage/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import RegisterForm from '../../components/Forms/RegistrationForm'
+import RegisterForm from './(components)/RegistrationForms/RegistrationForm'
 import UnLoggedHeader from '../../components/Headers/UnLoggedHeader'
 import { useContextWrap } from '../../contextAPI/context'
